// src/angular-cesium-entities-editor/models/polyline-edit-update.ts
import { EditPoint } from './edit-point';
import { Cartesian3 } from '../../angular-cesium/models/cartesian3';
import { BasicEditUpdate } from './basic-edit-update';
import { PolylineEditOptions } from './polyline-edit-options';

export interface PolylineEditUpdate extends BasicEditUpdate {
	positions?: Cartesian3[];
	updatedPosition?: Cartesian3;
	draggedPosition?: Cartesian3;
	points?: EditPoint[];
	updatedPoint?: EditPoint;
	polylineOptions?: PolylineEditOptions
}